Extract worker logic into @alxm/cf-worker; consume from both sites (az8.3)
Both sites duplicated a near-identical Cloudflare Worker (MIME map, R2
proxy with range support, TDM-Reservation header), with alxm.me adding
RSS handling. Hoist the shared logic into a new internal workspace
package, packages/cf-worker (@alxm/cf-worker), exposing named helpers
plus createFetchHandler({ rss }).

Each site's worker/index.js is now a thin entry:
  export default createFetchHandler({ rss: true|false })
alxm.me passes rss:true (serves env.RSS_PATH); alexmarshalltherapy.com
passes rss:false. R2 path prefixes still come from env.R2_PATHS at
request time, so no per-site parameterisation is needed there.

The worker unit tests (mime, r2, rss, plus a createFetchHandler
integration test covering both rss states) move into the package and
run via a new root `pnpm test:packages` script. Removes the duplicated
worker-mocks helper and the per-site worker tests.

Validated: 52 package tests pass; both site suites still green;
`wrangler deploy --dry-run` bundles each worker through the workspace
symlink (5.12 KiB each).
