+++ sites/alexmarshalltherapy.com/package.json
@@ -34,6 +34,7 @@
     "@11ty/eleventy": "^3.1.6",
     "@11ty/eleventy-img": "^6.0.4",
     "@11ty/eleventy-plugin-directory-output": "^1.0.2",
+    "@alxm/cf-worker": "workspace:*",
     "dotenv": "^17.4.2",
     "eleventy-plugin-cloudflare-kv": "^1.0.6",
     "eleventy-plugin-purgecss": "^0.6.0",
