+++ sites/alxm.me/package.json
@@ -37,6 +37,7 @@
     "@11ty/eleventy-img": "^6.0.4",
     "@11ty/eleventy-plugin-directory-output": "^1.0.2",
     "@11ty/eleventy-plugin-syntaxhighlight": "^5.0.2",
+    "@alxm/cf-worker": "workspace:*",
     "dotenv": "^17.4.2",
     "eleventy-plugin-cloudflare-kv": "^1.0.6",
     "eleventy-plugin-og-image": "^4.2.1",
