--- sites/alxm.me/tests/helpers/worker-mocks.js
@@ -1,7 +1,0 @@
-/**
- * Shared mock helpers for worker tests.
- */
-
-export function createRequest(url, headers = {}) {
-  return new Request(url, { headers });
-}
